+++ src/app/(auth)/_components/Register/UploadProfile.tsx
@@ -5,7 +5,7 @@ import { toast } from 'react-toastify';
 import { AddImageSVG, TrashcanSVG } from '@/../public/icons/svg';
 import { postProfileImage } from '@/lib/apis/userApi';
 import { useUserStore } from '@/store';
-import Button from '@/components/Button/Button';
+import { Button } from '@/components/Button';
 
 const UploadProfile = () => {
   const [imgSrc, setImgSrc] = useState<string | null>(null);
