+++ src/app/class/[id]/edit/_components/EditButtons.tsx
@@ -1,6 +1,5 @@
 import { UseFormHandleSubmit, FieldValues } from 'react-hook-form';
-import Button from '@/components/Button/Button';
-import UniqueButton from '@/components/Button/UniqueButton';
+import { Button, UniqueButton } from '@/components/Button';
 
 interface EditButtonsProps {
   handleEditCancel: () => void;
