+++ src/app/mypage/instructor/manage/myclass/[id]/_components/EditDayOff.tsx
@@ -1,10 +1,9 @@
 import { compareAsc } from 'date-fns';
 import { useState, useMemo, useRef, useEffect } from 'react';
-import Button from '@/components/Button/Button';
-import UniqueButton from '@/components/Button/UniqueButton';
-import DayOffCalendar from '@/components/Calendar/BasicCalendar';
 import { formatDateWithDay } from '@/utils/dateTimeUtils';
 import { getUniqueDates, calculateUnSelectedDate } from '@/utils/parseUtils';
+import { Button, UniqueButton } from '@/components/Button';
+import DayOffCalendar from '@/components/Calendar/BasicCalendar';
 
 /* eslint-disable no-unused-vars */
 interface EditDayOffProps {
