+++ src/app/mypage/user/review/@writeReviewModal/(.)writeReviewModal/_components/WriteReview.tsx
@@ -4,7 +4,7 @@ import { toast } from 'react-toastify';
 import { writeReview } from '@/lib/apis/reviewApis';
 import { accessTokenReissuance } from '@/lib/apis/userApi';
 import ClassSelect from './ClassSelect';
-import Button from '@/components/Button/Button';
+import { Button } from '@/components/Button';
 import Rating from '@/components/Review/Rating';
 import { SelectClassType, WriteReviewData } from '@/types/review';
 import { FetchError } from '@/types/types';
